+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Add support for extension 17513 and repo cleanup
 ### Enhancements
 - Return Json body instead of string if response is type Json 🚨 BREAKING CHANGE (users no longer need to `JSON.parse` the response request body)
 - Changed basic api to not require CycleTLS `config`
+- Add Cookie examples in readme and update Cookie API
 ### Bug Fixes
 - [Added support for TLS Certificate Compression (causing cloudflare handshake issues)](https://github.com/Danny-Dasilva/CycleTLS/issues/103)
 - [Send Correct Frame headers](https://github.com/Danny-Dasilva/CycleTLS/issues/101)